*Package structure refactoring
