+++ orchestrator/src/main/java/ch/unisg/airqueue/orchestrator/adapters/BookingDoneAdapter.java
@@ -1,7 +1,7 @@
 package ch.unisg.airqueue.orchestrator.adapters;
 
 import ch.unisg.airqueue.orchestrator.domain.Booking;
-import ch.unisg.airqueue.orchestrator.messages.BookingDoneEvent;
+import ch.unisg.airqueue.orchestrator.events.BookingDoneEvent;
 import ch.unisg.airqueue.orchestrator.messages.Message;
 import ch.unisg.airqueue.orchestrator.messages.MessageSender;
 import ch.unisg.airqueue.orchestrator.persistence.BookingRepository;
