+++ orchestrator/src/main/java/ch/unisg/airqueue/orchestrator/adapters/NotificationAdapter.java
@@ -3,7 +3,7 @@ package ch.unisg.airqueue.orchestrator.adapters;
 import ch.unisg.airqueue.orchestrator.domain.Booking;
 import ch.unisg.airqueue.orchestrator.messages.Message;
 import ch.unisg.airqueue.orchestrator.messages.MessageSender;
-import ch.unisg.airqueue.orchestrator.messages.NotificationCommand;
+import ch.unisg.airqueue.orchestrator.commands.NotificationCommand;
 import ch.unisg.airqueue.orchestrator.persistence.BookingRepository;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
