+++ orchestrator/src/main/java/ch/unisg/airqueue/orchestrator/commands/NotificationCommand.java
@@ -1,4 +1,4 @@
-package ch.unisg.airqueue.orchestrator.messages;
+package ch.unisg.airqueue.orchestrator.commands;
 
 public class NotificationCommand {
     private String content;
